C++-clean changes from Saurik
git-svn-id: https://svn.nmap.org/trunk/nmap@795 e0a8ed71-7df4-0310-8962-fdc924857419

+++ protocols.c
@@ -31,7 +31,7 @@ static int nmap_protocols_init() {
   while(fgets(line, sizeof(line), fp)) {
     lineno++;
     p = line;
-    while(*p && isspace(*p))
+    while(*p && isspace((int) *p))
       p++;
     if (*p == '#')
       continue;
